added function which evaluates the model against a baseline model and outputs the metrics we want to track. Also hidden away/cleaned up  evaluation scripts and moved most functions to the utils file
