+++ src/arcsf/eval/evaluate_model.py
@@ -1,0 +1,82 @@
+import numpy as np
+import torch
+import transformers
+
+from arcsf.data.data_module import EvalQADataset, get_data, qa_formatter_autoregression
+from arcsf.eval.scripts.analysis import get_metrics
+from arcsf.eval.utils import all_eval, combine_dicts
+from arcsf.utils import get_device
+
+RAND = 42
+
+
+def evaluate_model(
+    model: torch.nn.Module,
+    base_truth_ratios_path: str,
+    tokenizer: transformers.AutoTokenizer,
+    experiment_config: dict,
+) -> dict[float, float, float, float, float]:
+    """
+    Evaluates the model against a baseline model, given the model, tokenizer, experiment
+    config and the path to the baseline model truth ratios.
+
+    Args:
+        model: Model to perform the evaluation on
+        base_truth_ratios_path: path for the baseline mode truth ratios
+        tokenizer: tokenizer being used for the model
+        experiment_config: experiment config (currently a dictionary)
+
+    Returns:
+        dictionary : aggregated metrics for evaluation (generated using `get_metrics`)
+    """
+    # get datasets
+    forget_data, retain_data = get_data(
+        "tofu", random_seed=RAND, **experiment_config["data_config"]
+    )
+    # get available device
+    device = get_device()
+
+    # create datasets
+    RetainDataset = EvalQADataset(
+        retain_data,
+        tokenizer,
+        qa_formatter_autoregression,
+        "standard",
+        quantitative_eval=True,
+        qualitative_eval=True,
+        device=device,
+        random_seed=RAND,
+    )
+    ForgetDataset = EvalQADataset(
+        forget_data,
+        tokenizer,
+        qa_formatter_autoregression,
+        "standard",
+        quantitative_eval=True,
+        qualitative_eval=True,
+        device=device,
+        random_seed=RAND,
+    )
+
+    # perform evaluation on models to get the raw values
+    retain_values = all_eval(
+        model,
+        RetainDataset,
+        1,
+        device,
+        tokenizer,
+    )
+    forget_values = all_eval(
+        model,
+        ForgetDataset,
+        1,
+        device,
+        tokenizer,
+    )
+
+    # combine dictionaries and load the baseline model truth ratios
+    analysis_vals = combine_dicts(forget_values, retain_values)
+    base_truth_ratios = torch.tensor(np.loadtxt(base_truth_ratios_path))
+
+    # return the aggregated metrics using the get metrics function
+    return get_metrics(base_truth_ratios, analysis_vals)
