+++ tests/data/dummy_truth_ratios.txt
@@ -1,0 +1,3 @@
+1.498633323921395931e+01
+3.836439550588737282e+00
+7.477211974606858348e+00
